refactor category provider
